+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Support `Float64` (OME `double`) datatype by casting array data to `Float32`. 
 - `onHover` prop for `VivViewer`, `PictureInPictureViewer`, `SideBySideViewer` for deck.gl callback.
-
+- Add `handleCoordnate` `hoverHook` for PIP.
 ### Changed
 
 ## 0.9.2